refactor code to host_initiator
