+++ config/api.yml
@@ -1880,6 +1880,30 @@
         :identifier: host_aggregate_tag
       - :name: unassign
         :identifier: host_aggregate_tag
+  :host_initiators:
+    :description: Host Initiators
+    :identifier: host_initiator
+    :options:
+    - :collection
+    :verbs: *gp
+    :klass: HostInitiator
+    :subcollections:
+    :collection_actions:
+      :get:
+      - :name: read
+        :identifier: host_initiator_show_list
+      :post:
+      - :name: create
+        :identifier: host_initiator_new
+      - :name: refresh
+        :identifier: host_initiator_refresh
+    :resource_actions:
+      :get:
+      - :name: read
+        :identifier: host_initiator_show
+      :post:
+      - :name: refresh
+        :identifier: host_initiator_refresh
   :hosts:
     :description: Hosts
     :identifier: host
@@ -2474,30 +2498,6 @@
         :identifier: physical_storage_refresh
       - :name: delete
         :identifier: physical_storage_delete
-  :physical_storage_consumers:
-    :description: Physical Storage Consumers
-    :identifier: physical_storage_consumer
-    :options:
-      - :collection
-    :verbs: *gp
-    :klass: PhysicalStorageConsumer
-    :subcollections:
-    :collection_actions:
-      :get:
-        - :name: read
-          :identifier: physical_storage_consumer_show_list
-      :post:
-        - :name: create
-          :identifier: physical_storage_consumer_new
-        - :name: refresh
-          :identifier: physical_storage_consumer_refresh
-    :resource_actions:
-      :get:
-        - :name: read
-          :identifier: physical_storage_consumer_show
-      :post:
-        - :name: refresh
-          :identifier: physical_storage_consumer_refresh
   :physical_switches:
     :description: Physical Switches
     :identifier: physical_switch
